requirements.txt
opencv-python==4.10.0.84
numpy==1.24.3
tensorflow==2.13.1
pillow==10.4.0
python-dotenv==1.0.1
flask==3.0.3
ultralytics==8.2.77